Revert back to a dumber collision system, as the precise one was too complicated for the HInt

+++ src/physics.c
@@ -14,8 +14,6 @@ PlayerStat * bl_stat;
 PlayerStat * gr_stat;
 
 u8 collided;
-u8 blue_collided;
-u8 green_collided;
 
 //TODO: Organize. Make more understandable
 
@@ -752,20 +750,15 @@ static inline void class_tree(){
             nastie_tree();
             if(curr->status == DEAD)
                 return;
-            if(blue_collided){
-                if(ACT_collision(blue_player, curr)){
+            if(collided){
+                if(blue_player && ACT_collision(blue_player, curr)){
                     kill(blue_player, 0, -2*FALLSPEED);
                     blue_player = NULL;
                 }
-            }
-            if(green_collided){
-                if(ACT_collision(green_player, curr)){
+                if(green_player && ACT_collision(green_player, curr)){
                     kill(green_player, 0, -2*FALLSPEED);
                     green_player = NULL;
                 }
-            }
-            if(collided){
-
                 Actor * proj_act = ACT_getFirst(&pl_projectiles);
                 while(proj_act){
                     if(ACT_collision(proj_act, curr)){
@@ -792,24 +785,13 @@ static inline void class_tree(){
 
 void PHY_HCallback(){
     if(GET_VDPSTATUS(VDP_SPRCOLLISION_FLAG)){
-        register int line = VDP_getHIntCounter();
-        if(line < player_lines[B1] && line > player_lines[B0]){
-            blue_collided = TRUE;
-        }
-        if(line < player_lines[G1] && line > player_lines[G0]){
-            green_collided = TRUE;
-        }
-        else {
-            collided = TRUE;
-        }
+        collided = TRUE;    
     }
 }
 
 u8 PHY_init(Board * board, PlayerStat * bl_stats, PlayerStat * gr_stats){
     env = board;
     collided = FALSE;
-    blue_collided = FALSE;
-    green_collided = FALSE;
     SYS_setHIntCallback(PHY_HCallback);
     VDP_setHIntCounter(0);
     VDP_setHInterrupt(1);
@@ -825,8 +807,6 @@ u8 PHY_init(Board * board, PlayerStat * bl_stats, PlayerStat * gr_stats){
 
 void PHY_end(){
     collided = FALSE;
-    blue_collided = FALSE;
-    green_collided = FALSE;
     VDP_setHInterrupt(0);
     ACT_end(&nasties);
     ACT_end(&fx_buf);
@@ -858,20 +838,10 @@ u8 PHY_computeStatus(Actor * actor){
 
 
 void PHY_update(){
-    if(blue_player){
-        player_lines[B0] = POS_TO_PX(blue_player->pos[Y]) - blue_player->character->spr_pos[Y] + BOARD_OFFSET_Y;
-        player_lines[B1] = player_lines[B0] + 15;
-    }
-    if(green_player){
-        player_lines[G0] = POS_TO_PX(green_player->pos[Y]) - green_player->character->spr_pos[Y] + BOARD_OFFSET_Y;
-        player_lines[G1] = player_lines[G0] + 15;
-    }
     ACT_update(&players);
     ACT_update(&nasties);
     ACT_update(&projectiles);
     ACT_update(&pl_projectiles);
     ACT_update(&fx_buf);
     if(collided) collided = FALSE;
-    if(blue_collided) blue_collided = FALSE;
-    if(green_collided) green_collided = FALSE;
 }
